requirements.txt
spotipy>=2.22.1
pytest>=7.4.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.18.0